Implement wait queue (with bugs)

+++ lab5/include/fork.h
@@ -1,6 +1,6 @@
 #ifndef __FORK_H_
 #define __FORK_H_
 
-int do_fork(struct pt_regs *regs);
+size_t do_fork(struct pt_regs *regs);
 
 #endif
